fix: fix data type of columns

+++ src/main/java/com/greenatom/domain/entity/Employee.java
@@ -41,7 +41,7 @@ public class Employee {
     private JobPosition jobPosition;
 
     @Column(name = "salary")
-    private Integer salary;
+    private Long salary;
 
     @Column(name = "email")
     private String email;
